Site rebuild for use with the new Vue module for the icons
Flippin' Javascript npm stuff. Such a headache and waste of hours. Had to rebuild the whole site to use Vue 3 to get things working at all. So here we go.

Currently relies on the locally prepublished icons Vue module. Need to switch this to the published one once it's live.